move queries to lib
