+++ project/apps/bff/src/app/app.constant.ts
@@ -20,21 +20,3 @@ export const AppPath = {
   Repost: 'repost',
   Like: 'change-like-status',
 } as const;
-
-export const DefaultPostsLimit = {
-  Query: 25,
-  Search: 20,
-} as const;
-
-export const DefaultSortParam = {
-  Direction: 'desc',
-  Type: 'postedDate',
-} as const;
-
-export const TagDefaultParam = {
-  MinLength: 3,
-  MaxLength: 10,
-  Amount: 8,
-};
-
-export const DEFAULT_COMMENTS_LIMIT = 50;
