+++ project/libs/shared/shared-queries/README.md
@@ -1,0 +1,7 @@
+# shared-shared-queries
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test shared-shared-queries` to execute the unit tests via [Jest](https://jestjs.io).
