+++ project/libs/shared/shared-queries/src/index.ts
@@ -1,0 +1,3 @@
+export * from './lib/comment.query';
+export * from './lib/post.query';
+export * from './lib/search.query';
