+++ project/libs/shared/shared-queries/src/lib/comment.query.ts
@@ -1,6 +1,6 @@
 import { IsNumber,IsOptional } from 'class-validator';
 import { Transform } from 'class-transformer';
-import { DEFAULT_COMMENTS_LIMIT } from '../comments/comments.constant';
+import { DEFAULT_COMMENTS_LIMIT } from './query.constant';
 
 export class CommentQuery {
   @Transform(({ value } ) => +value || DEFAULT_COMMENTS_LIMIT)
